apps/image-classifier-app/python-models/requirements.txt
Flask==2.2.2
Flask-Cors==5.0.0
Werkzeug==2.2.2
matplotlib==3.9.2
numpy>=1.23,<2.0
pandas==2.2.3
Pillow==11.0.0
scikit_learn==1.3.2
seaborn==0.13.2
protobuf==3.20.3
tensorflow==2.17.0
timm==1.0.11
torch==2.3.0
torchvision==0.18.0